+++ src/PortalBase/Builder/Component/HttpHandler.php
@@ -54,6 +54,41 @@ class HttpHandler extends HttpHandlerContract
         $this->deleteMethod = $delete instanceof \Closure ? new SerializableClosure($delete) : null;
     }
 
+    public function getRunMethod(): ?\Closure
+    {
+        return $this->runMethod instanceof SerializableClosure ? $this->runMethod->getClosure() : null;
+    }
+
+    public function getOptionsMethod(): ?\Closure
+    {
+        return $this->optionsMethod instanceof SerializableClosure ? $this->optionsMethod->getClosure() : null;
+    }
+
+    public function getGetMethod(): ?\Closure
+    {
+        return $this->getMethod instanceof SerializableClosure ? $this->getMethod->getClosure() : null;
+    }
+
+    public function getPostMethod(): ?\Closure
+    {
+        return $this->postMethod instanceof SerializableClosure ? $this->postMethod->getClosure() : null;
+    }
+
+    public function getPatchMethod(): ?\Closure
+    {
+        return $this->patchMethod instanceof SerializableClosure ? $this->patchMethod->getClosure() : null;
+    }
+
+    public function getPutMethod(): ?\Closure
+    {
+        return $this->putMethod instanceof SerializableClosure ? $this->putMethod->getClosure() : null;
+    }
+
+    public function getDeleteMethod(): ?\Closure
+    {
+        return $this->deleteMethod instanceof SerializableClosure ? $this->deleteMethod->getClosure() : null;
+    }
+
     protected function supports(): string
     {
         return $this->path;