+++ src/PortalBase/Builder/Component/StatusReporter.php
@@ -28,6 +28,11 @@ class StatusReporter extends StatusReporterContract
         $this->runMethod = $run instanceof \Closure ? new SerializableClosure($run) : null;
     }
 
+    public function getRunMethod(): ?\Closure
+    {
+        return $this->runMethod instanceof SerializableClosure ? $this->runMethod->getClosure() : null;
+    }
+
     public function supportsTopic(): string
     {
         return $this->topic;