+++ src/PortalBase/FlowComponent/Exception/CodeOriginNotFound.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Heptacom\HeptaConnect\Portal\Base\FlowComponent\Exception;
+
+class CodeOriginNotFound extends \RuntimeException
+{
+    private object $flowComponent;
+
+    public function __construct(object $flowComponent, int $code, ?\Throwable $throwable = null)
+    {
+        parent::__construct('Could not find code origin of a flow component', $code, $throwable);
+        $this->flowComponent = $flowComponent;
+    }
+
+    public function getFlowComponent(): object
+    {
+        return $this->flowComponent;
+    }
+}