+++ src/PortalBase/StatusReporting/StatusReporterStack.php
@@ -7,6 +7,7 @@ namespace Heptacom\HeptaConnect\Portal\Base\StatusReporting;
 use Heptacom\HeptaConnect\Portal\Base\StatusReporting\Contract\StatusReporterContract;
 use Heptacom\HeptaConnect\Portal\Base\StatusReporting\Contract\StatusReporterStackInterface;
 use Heptacom\HeptaConnect\Portal\Base\StatusReporting\Contract\StatusReportingContextInterface;
+use Psr\Log\LoggerInterface;
 
 class StatusReporterStack implements StatusReporterStackInterface
 {
@@ -15,14 +16,17 @@ class StatusReporterStack implements StatusReporterStackInterface
      */
     private array $statusReporters;
 
+    private LoggerInterface $logger;
+
     /**
      * @param iterable<array-key, StatusReporterContract> $statusReporters
      */
-    public function __construct(iterable $statusReporters)
+    public function __construct(iterable $statusReporters, LoggerInterface $logger)
     {
         /** @var StatusReporterContract[] $rewindableStatusReporters */
         $rewindableStatusReporters = \iterable_to_array($statusReporters);
         $this->statusReporters = $rewindableStatusReporters;
+        $this->logger = $logger;
     }
 
     public function next(StatusReportingContextInterface $context): array
@@ -32,6 +36,10 @@ class StatusReporterStack implements StatusReporterStackInterface
         if (!$statusReporter instanceof StatusReporterContract) {
             return [];
         }
+
+        $this->logger->debug('Execute FlowComponent status reporter', [
+            'status_reporter' => $statusReporter,
+        ]);
 
         return $statusReporter->report($context, $this);
     }